finished triangle fill and painter's algorithm
Each face now gets filled in order from furthest to closest, and is
clipped to the near frustum plane. The rest of the clipping is done in
2d during the triangle rasterization.

The texture index in each face determines the colour used to fill in
each face.

Instead of using the normal vector given in the map file, it is
calculated from the 3 points in the plane, and will face towards or away
from the origin based on the type (0 is towards the origin, 1 is away
from the origin). This makes the mapping process much easier.