finished the pillar map, added a settings file
Some settings are now in a text file, so they can be changed without
recompiling the source code. The depth ordering was changed slightly
(the midpoint is used instead of the median, but there are still some
visual bugs present). Frames per second is now displayed, and the delay
time is decreased if the time taken in the last loop was too long, to
keep the loop interval constant.