changed rendering settings, lines now are clipped
SDL_RENDER_ACCELERATED was causing low framerate for some reason, but
using software rendering fixed this. The Cohen Sutherland algorithm is
being used to clip lines to the window area because software rendering
just truncates the x or y value to the screen boundaries. Setting the
draw edges value to 2 in the settings will now draw the line which
divides the quadrilateral created after clipping a triangle with the
near plane, in gray.